backend/pyproject.toml
[tool.poetry]
name = "tgbusiness-backend"
version = "1.0.0"
description = "TG Business Suite Backend"
authors = ["TG Business Suite"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
gunicorn = "^21.2.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
telethon = "^1.35.0"
celery = {extras = ["redis"], version = "^5.3.6"}
stripe = "^7.10.0"
aiohttp = "^3.9.1"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
email-validator = "^2.1.0.post1"
redis = "^5.0.1"
starlette = "^0.35.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
